added missing import and refactored to new datatypes

+++ backend/src/main/java/edu/alexu/cse/dripmeup/dto/ItemDTO.java
@@ -10,7 +10,7 @@ public class ItemDTO {
     private String ProductName;
     private String ProductVariantSize;
     private String ProductVariantColor;
-    private float ProductVariantPrice;
+    private double ProductVariantPrice;
     private Integer ProductVariantQuantity;
-    private float ItemTotalPrice;
+    private double ItemTotalPrice;
 }
